address the PR comments
